[NVIDIA] Extract ITopologyRunner into the separate header
