+++ modules/nvidia_plugin/src/cuda_eager_topology_runner.hpp
@@ -1,20 +1,14 @@
-// Copyright (C) 2018-2021 Intel Corporation
+// Copyright (C) 2018-2023 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 //
 
 #pragma once
 
 #include <ops/subgraph.hpp>
+#include "cuda_itopology_runner.hpp"
 
 namespace ov {
 namespace nvidia_gpu {
-
-struct ITopologyRunner {
-    virtual void Run(const InferenceRequestContext& context, const DeviceMemBlock& memoryBlock) const = 0;
-    virtual void UpdateContext(InferenceRequestContext& context, const DeviceMemBlock& memoryBlock) const = 0;
-    virtual const SubGraph& GetSubGraph() const = 0;
-    virtual ~ITopologyRunner() = default;
-};
 
 class EagerTopologyRunner final : public SubGraph, public ITopologyRunner {
 public:
